fix change eq

+++ backend/HUAWEI/views/eq_views.py
@@ -32,7 +32,8 @@ class EquipmentView(APIView):
 
         # 在数据库中更新设备
         new_site = Site.objects.get(site_id = thesite.site_id)
-        Equipment.objects.filter(site = thesite.pk).delete()
+        if Equipment.objects.filter(site = thesite.pk):
+            Equipment.objects.filter(site = thesite.pk).delete()
         for i in range(0, eq_num):
             eq_data = {
                 'eq_id': str(new_site.pk) + str(i),
